+++ .gitignore
@@ -20,3 +20,6 @@ build
 # other
 eclipse
 run
+
+# Files from Forge MDK
+forge*changelog.txt
