--- src/main/java/io/alwa/featuredserver/ServerDataHelper.java
@@ -1,7 +1,0 @@
-package io.alwa.featuredserver;
-
-public class ServerDataHelper {
-
-    public String serverName;
-    public String serverIP;
-}
